feat: 迁移项目到 xmake
- feat: 链接库全部迁移到 `xmake`
- feat: xml 支持 `utf-8`, `utf-16le`, `utf-16be` 编码混合
- fix: `DrawTextA` 渲染 `utf-8` 文件乱码 `code page` 设置了也无效(MessageBox能正常显示)
- feat: `dpi` 支持，初始无需调用代码，设置 dpi 感知即可
- feat: 拆分控件到出去
- feat: mingw 适配
- fix: 修复 `mingw` 的 `dpi` 窗口缩放问题

+++ 3rd/CxImage/ximapal.cpp
@@ -444,7 +444,7 @@ void CxImage::SetPalette(uint32_t n, uint8_t *r, uint8_t *g, uint8_t *b)
 	if (!g) g = r;
 	if (!b) b = g;
 	RGBQUAD* ppal=GetPalette();
-	uint32_t m=min(n,head.biClrUsed);
+	uint32_t m=min(n,(uint32_t)(head.biClrUsed));
 	for (uint32_t i=0; i<m;i++){
 		ppal[i].rgbRed=r[i];
 		ppal[i].rgbGreen=g[i];
@@ -457,7 +457,7 @@ void CxImage::SetPalette(rgb_color *rgb,uint32_t nColors)
 {
 	if ((!rgb)||(pDib==NULL)||(head.biClrUsed==0)) return;
 	RGBQUAD* ppal=GetPalette();
-	uint32_t m=min(nColors,head.biClrUsed);
+	uint32_t m=min(nColors,(uint32_t)(head.biClrUsed));
 	for (uint32_t i=0; i<m;i++){
 		ppal[i].rgbRed=rgb[i].r;
 		ppal[i].rgbGreen=rgb[i].g;
@@ -469,7 +469,7 @@ void CxImage::SetPalette(rgb_color *rgb,uint32_t nColors)
 void CxImage::SetPalette(RGBQUAD* pPal,uint32_t nColors)
 {
 	if ((pPal==NULL)||(pDib==NULL)||(head.biClrUsed==0)) return;
-	memcpy(GetPalette(),pPal,min(GetPaletteSize(),nColors*sizeof(RGBQUAD)));
+	memcpy(GetPalette(),pPal,min(GetPaletteSize(),(uint32_t)(nColors*sizeof(RGBQUAD))));
 	info.last_c_isvalid = false;
 }
 ////////////////////////////////////////////////////////////////////////////////
@@ -654,10 +654,10 @@ void CxImage::SetClrImportant(uint32_t ncolors)
 
 	switch(head.biBitCount){
 	case 1:
-		head.biClrImportant = min(ncolors,2);
+		head.biClrImportant = min(ncolors,(uint32_t)2);
 		break;
 	case 4:
-		head.biClrImportant = min(ncolors,16);
+		head.biClrImportant = min(ncolors,(uint32_t)16);
 		break;
 	case 8:
 		head.biClrImportant = ncolors;
